bfd: Set the thumb (16bit) flag for code sections in the arm-mingw32 target
Also don't make code sections writeable.

In this target, all code is supposed to be in thumb mode.
Ideally, the thumb flag should be set based on what format
the functions in the section are.

This is required for the object files to work properly when
linked into an app by the ms linker.