bfd: Add a new target named arm-*-mingw32
This is for targeting "normal" windows on arm, such as Windows RT
or Windows Phone 8, both which technically are "normal" windows,
in contrary to Windows CE.

This shares much code with the existing ARM_WINCE conditionals
(which mostly are generic to microsoft's PE/COFF for ARM and
not specific to windows CE). The main difference to the old
CE target is that this target uses a separate COFF magic value
(ARMV7PEMAGIC) which is documented as "ARMv7 (or higher) Thumb
mode only".